Whole 6.2: Number validations

+++ pkg/schemas/model.go
@@ -91,7 +91,7 @@ type Type struct {
 	Version string `json:"$schema,omitempty"` // section 6.1
 	Ref     string `json:"$ref,omitempty"`    // section 7
 	// RFC draft-wright-json-schema-validation-00, section 5
-	MultipleOf           int              `json:"multipleOf,omitempty"`           // section 5.1
+	MultipleOf           float64          `json:"multipleOf,omitempty"`           // section 5.1
 	Maximum              float64          `json:"maximum,omitempty"`              // section 5.2
 	ExclusiveMaximum     float64          `json:"exclusiveMaximum,omitempty"`     // section 5.3
 	Minimum              float64          `json:"minimum,omitempty"`              // section 5.4
